# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Error {
  error: string;
}

export type ItemType = typeof ItemType[keyof typeof ItemType];


export const ItemType = {
  lost: 'lost',
  found: 'found',
} as const;

export type ItemStatus = typeof ItemStatus[keyof typeof ItemStatus];


export const ItemStatus = {
  LOST: 'LOST',
  FOUND: 'FOUND',
  MATCHED: 'MATCHED',
  VERIFICATION_PENDING: 'VERIFICATION_PENDING',
  VERIFIED: 'VERIFIED',
  RESOLVED: 'RESOLVED',
} as const;

export interface Item {
  id: string;
  type: ItemType;
  category: string;
  name: string;
  brand: string;
  model?: string;
  color: string;
  location: string;
  date: string;
  time: string;
  description: string;
  imageUrl: string;
  status: ItemStatus;
  createdAt: string;
  ownerName?: string;
  privateClueCount?: number;
}

export type ItemInputType = typeof ItemInputType[keyof typeof ItemInputType];


export const ItemInputType = {
  lost: 'lost',
  found: 'found',
} as const;

export interface ItemInput {
  type: ItemInputType;
  category: string;
  /** @minLength 1 */
  name: string;
  brand: string;
  model?: string;
  color: string;
  location: string;
  date: string;
  time: string;
  description: string;
  imageUrl?: string;
  privateClues?: string[];
}

export type MatchStatus = typeof MatchStatus[keyof typeof MatchStatus];


export const MatchStatus = {
  POTENTIAL: 'POTENTIAL',
  VERIFICATION_PENDING: 'VERIFICATION_PENDING',
  VERIFIED: 'VERIFIED',
  RESOLVED: 'RESOLVED',
} as const;

export interface Match {
  id: string;
  lostItem: Item;
  foundItem: Item;
  similarity: number;
  factors: string[];
  status: MatchStatus;
}

export interface MatchInput {
  lostItemId: string;
  foundItemId: string;
}

export type ClaimVerificationStatus = typeof ClaimVerificationStatus[keyof typeof ClaimVerificationStatus];


export const ClaimVerificationStatus = {
  PENDING: 'PENDING',
  PASSED: 'PASSED',
  FAILED: 'FAILED',
  ADDITIONAL_REVIEW: 'ADDITIONAL_REVIEW',
} as const;

export type ClaimIdentityStatus = typeof ClaimIdentityStatus[keyof typeof ClaimIdentityStatus];


export const ClaimIdentityStatus = {
  VERIFIED: 'VERIFIED',
  UNVERIFIED: 'UNVERIFIED',
} as const;

export interface Claim {
  id: string;
  item: Item;
  claimantName: string;
  verificationStatus: ClaimVerificationStatus;
  identityStatus: ClaimIdentityStatus;
  evidenceMatched?: string[];
  createdAt: string;
}

export interface ClaimInput {
  itemId: string;
  evidence: string;
  identityEmail?: string;
}

export interface VerificationInput {
  /** @minLength 1 */
  evidence: string;
}

export type VerificationResultStatus = typeof VerificationResultStatus[keyof typeof VerificationResultStatus];


export const VerificationResultStatus = {
  PASSED: 'PASSED',
  FAILED: 'FAILED',
  ADDITIONAL_REVIEW: 'ADDITIONAL_REVIEW',
} as const;

export interface VerificationResult {
  status: VerificationResultStatus;
  message: string;
  evidenceMatched: string[];
}

export type HandoverInputAction = typeof HandoverInputAction[keyof typeof HandoverInputAction];


export const HandoverInputAction = {
  ARRANGE: 'ARRANGE',
  CONFIRM: 'CONFIRM',
} as const;

export interface HandoverInput {
  itemId: string;
  action: HandoverInputAction;
  note?: string;
}

export type HandoverStatus = typeof HandoverStatus[keyof typeof HandoverStatus];


export const HandoverStatus = {
  ARRANGED: 'ARRANGED',
  COMPLETED: 'COMPLETED',
} as const;

export interface Handover {
  id: string;
  item: Item;
  status: HandoverStatus;
  nextStep: string;
}

export type DashboardStats = {
  reported: number;
  matches: number;
  pendingClaims: number;
  returned: number;
};

export type DashboardRecentActivityItem = {
  id: string;
  title: string;
  description: string;
  timestamp: string;
  type: string;
};

export interface Dashboard {
  stats: DashboardStats;
  recentItems: Item[];
  recentActivity: DashboardRecentActivityItem[];
}

export type AdminMetricsTotals = {
  lost: number;
  found: number;
  matches: number;
  pending: number;
  returned: number;
};

export type AdminMetricsCategoriesItem = {
  category: string;
  count: number;
};

export type AdminMetricsTrendItem = {
  label: string;
  lost: number;
  found: number;
};

export interface AdminMetrics {
  totals: AdminMetricsTotals;
  categories: AdminMetricsCategoriesItem[];
  trend: AdminMetricsTrendItem[];
}

/**
 * Not found
 */
export type NotFoundResponse = Error;

export type ListItemsParams = {
type?: ListItemsType;
status?: string;
search?: string;
};

export type ListItemsType = typeof ListItemsType[keyof typeof ListItemsType];


export const ListItemsType = {
  lost: 'lost',
  found: 'found',
} as const;