# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimInput {
  itemId: string;
  evidence: string;
  identityEmail?: string;
}