# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardRecentActivityItem } from './dashboardRecentActivityItem';
import type { DashboardStats } from './dashboardStats';
import type { Item } from './item';

export interface Dashboard {
  stats: DashboardStats;
  recentItems: Item[];
  recentActivity: DashboardRecentActivityItem[];
}