# lib/api-zod/src/generated/types/handoverInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
import type { HandoverInputAction } from './handoverInputAction';

export interface HandoverInput {
  itemId: string;
  action: HandoverInputAction;
  note?: string;
}