# lib/api-zod/src/generated/types/item.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemStatus } from './itemStatus';
import type { ItemType } from './itemType';

export interface Item {
  id: string;
  type: ItemType;
  category: string;
  name: string;
  brand: string;
  model?: string;
  color: string;
  location: string;
  date: string;
  time: string;
  description: string;
  imageUrl: string;
  status: ItemStatus;
  createdAt: string;
  ownerName?: string;
  privateClueCount?: number;
}