# lib/api-zod/src/generated/types/itemInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ItemInputType } from './itemInputType';

export interface ItemInput {
  type: ItemInputType;
  category: string;
  /** @minLength 1 */
  name: string;
  brand: string;
  model?: string;
  color: string;
  location: string;
  date: string;
  time: string;
  description: string;
  imageUrl?: string;
  privateClues?: string[];
}