# lib/api-zod/src/generated/types/matchInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MatchInput {
  lostItemId: string;
  foundItemId: string;
}