# lib/api-zod/src/generated/types/verificationResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * LostLink secure lost-and-found platform API
 * OpenAPI spec version: 0.1.0
 */
import type { VerificationResultStatus } from './verificationResultStatus';

export interface VerificationResult {
  status: VerificationResultStatus;
  message: string;
  evidenceMatched: string[];
}